Вынес типизацию пользователя в отдельный файл
