+++ packages/client/src/api/AuthAPI.ts
@@ -1,31 +1,5 @@
 import { API_URL } from './consts';
 
-export interface SignupData {
-  first_name: string;
-  second_name: string;
-  login: string;
-  email: string;
-  password: string;
-  phone: string;
-}
-
-export interface SigninData {
-  login: string;
-  password: string;
-}
-
-export interface User {
-  id: number;
-  first_name: string;
-  second_name: string;
-  display_name: string;
-  login: string;
-  email: string;
-  phone: string;
-  avatar: string;
-  role?: string;
-}
-
 type OptionsType = {
   headers: { 'Content-Type': string };
   credentials: RequestCredentials | undefined;
