+++ packages/client/src/user.d.ts
@@ -1,0 +1,25 @@
+interface SignupData {
+  first_name: string;
+  second_name: string;
+  login: string;
+  email: string;
+  password: string;
+  phone: string;
+}
+
+interface SigninData {
+  login: string;
+  password: string;
+}
+
+interface User {
+  id: number;
+  first_name: string;
+  second_name: string;
+  display_name: string;
+  login: string;
+  email: string;
+  phone: string;
+  avatar: string;
+  role?: string;
+}
